Call Management!
Need to work on life management, score management, and scene management. The score is defined by how far the obstacle moves; I need to figure out how to assign the obstacle tag onto the prefab. The hearts are not disappearing after 3 frames which means that I need to check the death variable within Launch on top of the reload function.